feat: moved scheduler functions to a specific module
