--- test.c
@@ -1,6 +1,0 @@
-
-
-int main (void){
-
-    return 0;
-}
